Refactor import paths for `get_patched_openai_client` in astra_assistants components
