+++ src/backend/base/langflow/components/astra_assistants/list_assistants.py
@@ -1,4 +1,4 @@
-from langflow.components.astra_assistants.util import get_patched_openai_client
+from langflow.base.astra_assistants.util import get_patched_openai_client
 from langflow.custom.custom_component.component_with_cache import ComponentWithCache
 from langflow.schema.message import Message
 from langflow.template.field.base import Output
